chore: migrate fetch lists to list repository
