+++ core/network/src/main/kotlin/com/divinelink/core/network/account/service/AccountService.kt
@@ -1,6 +1,5 @@
 package com.divinelink.core.network.account.service
 
-import com.divinelink.core.network.account.model.ListsResponse
 import com.divinelink.core.network.media.model.movie.MoviesResponseApi
 import com.divinelink.core.network.media.model.tv.TvResponseApi
 import kotlinx.coroutines.flow.Flow
@@ -34,9 +33,4 @@ interface AccountService {
     accountId: String,
     sessionId: String,
   ): Flow<TvResponseApi>
-
-  fun fetchUserLists(
-    accountId: String,
-    page: Int,
-  ): Flow<ListsResponse>
 }
