+++ core/network/src/main/kotlin/com/divinelink/core/network/list/service/ListService.kt
@@ -1,9 +1,11 @@
 package com.divinelink.core.network.list.service
 
+import com.divinelink.core.network.account.model.ListsResponse
 import com.divinelink.core.network.list.model.CreateListRequest
 import com.divinelink.core.network.list.model.CreateListResponse
 import com.divinelink.core.network.list.model.add.AddToListResponse
 import com.divinelink.core.network.list.model.details.ListDetailsResponse
+import kotlinx.coroutines.flow.Flow
 
 interface ListService {
   suspend fun addItemToList(
@@ -16,6 +18,11 @@ interface ListService {
     listId: Int,
     page: Int,
   ): Result<ListDetailsResponse>
+
+  fun fetchUserLists(
+    accountId: String,
+    page: Int,
+  ): Flow<ListsResponse>
 
   suspend fun createList(request: CreateListRequest): Result<CreateListResponse>
 }
