+++ core/network/src/main/kotlin/com/divinelink/core/network/list/service/ProdListService.kt
@@ -1,5 +1,7 @@
 package com.divinelink.core.network.list.service
 
+import com.divinelink.core.network.account.model.ListsResponse
+import com.divinelink.core.network.account.util.buildFetchListsUrl
 import com.divinelink.core.network.client.AuthTMDbClient
 import com.divinelink.core.network.list.model.CreateListRequest
 import com.divinelink.core.network.list.model.CreateListResponse
@@ -10,6 +12,8 @@ import com.divinelink.core.network.list.model.details.ListDetailsResponse
 import com.divinelink.core.network.list.util.buildAddItemsToListUrl
 import com.divinelink.core.network.list.util.buildFetchListDetailsUrl
 import com.divinelink.core.network.list.util.buildListUrl
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flow
 
 class ProdListService(private val client: AuthTMDbClient) : ListService {
 
@@ -41,6 +45,20 @@ class ProdListService(private val client: AuthTMDbClient) : ListService {
         page = page,
       ),
     )
+  }
+
+  override fun fetchUserLists(
+    accountId: String,
+    page: Int,
+  ): Flow<ListsResponse> {
+    val url = buildFetchListsUrl(
+      accountId = accountId,
+      page = page,
+    )
+
+    return flow {
+      emit(client.get<ListsResponse>(url = url))
+    }
   }
 
   override suspend fun createList(request: CreateListRequest): Result<CreateListResponse> =
