+++ core/testing/src/main/kotlin/com/divinelink/core/testing/repository/TestAccountRepository.kt
@@ -2,7 +2,6 @@ package com.divinelink.core.testing.repository
 
 import com.divinelink.core.data.account.AccountRepository
 import com.divinelink.core.model.PaginationData
-import com.divinelink.core.model.list.ListItem
 import com.divinelink.core.model.media.MediaItem
 import kotlinx.coroutines.flow.Flow
 import org.mockito.kotlin.any
@@ -53,12 +52,6 @@ class TestAccountRepository {
         accountId = any(),
         sessionId = any(),
       ),
-    ).thenReturn(response)
-  }
-
-  suspend fun mockFetchUserLists(response: Flow<Result<PaginationData<ListItem>>>) {
-    whenever(
-      mock.fetchUserLists(accountId = any(), page = any()),
     ).thenReturn(response)
   }
 }
